Refactor: Move bootstrap interface setup from sled-agent to sled-hardware
