+++ sled-agent/src/bootstrap/agent.rs
@@ -25,20 +25,19 @@ use crate::services::ServiceManager;
 use crate::sp::SpHandle;
 use crate::updates::UpdateManager;
 use futures::stream::{self, StreamExt, TryStreamExt};
-use illumos_utils::dladm::{
-    self, Dladm, Etherstub, EtherstubVnic, GetMacError, PhysicalLink,
-};
+use illumos_utils::dladm::{Dladm, Etherstub, EtherstubVnic, GetMacError};
 use illumos_utils::zfs::{
     self, Mountpoint, Zfs, ZONE_ZFS_DATASET, ZONE_ZFS_DATASET_MOUNTPOINT,
 };
 use illumos_utils::zone::Zones;
 use omicron_common::address::Ipv6Subnet;
 use omicron_common::address::SLED_PREFIX;
-use omicron_common::api::external::{Error as ExternalError, MacAddr};
+use omicron_common::api::external::Error as ExternalError;
 use omicron_common::backoff::{
     retry_notify, retry_policy_internal_service_aggressive, BackoffError,
 };
 use serde::{Deserialize, Serialize};
+use sled_hardware::underlay::bootstrap_ip;
 use sled_hardware::HardwareManager;
 use slog::Logger;
 use std::borrow::Cow;
@@ -48,12 +47,6 @@ use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use thiserror::Error;
 use tokio::sync::Mutex;
-
-/// Initial octet of IPv6 for bootstrap addresses.
-pub(crate) const BOOTSTRAP_PREFIX: u16 = 0xfdb0;
-
-/// IPv6 prefix mask for bootstrap addresses.
-pub(crate) const BOOTSTRAP_MASK: u8 = 64;
 
 /// Describes errors which may occur while operating the bootstrap service.
 #[derive(Error, Debug)]
@@ -204,32 +197,6 @@ fn get_sled_agent_request_path() -> PathBuf {
         .join("sled-agent-request.toml")
 }
 
-fn mac_to_bootstrap_ip(mac: MacAddr, interface_id: u64) -> Ipv6Addr {
-    let mac_bytes = mac.into_array();
-    assert_eq!(6, mac_bytes.len());
-
-    Ipv6Addr::new(
-        BOOTSTRAP_PREFIX,
-        ((mac_bytes[0] as u16) << 8) | mac_bytes[1] as u16,
-        ((mac_bytes[2] as u16) << 8) | mac_bytes[3] as u16,
-        ((mac_bytes[4] as u16) << 8) | mac_bytes[5] as u16,
-        (interface_id >> 48 & 0xffff).try_into().unwrap(),
-        (interface_id >> 32 & 0xffff).try_into().unwrap(),
-        (interface_id >> 16 & 0xffff).try_into().unwrap(),
-        (interface_id & 0xfff).try_into().unwrap(),
-    )
-}
-
-// TODO(https://github.com/oxidecomputer/omicron/issues/945): This address
-// could be randomly generated when it no longer needs to be durable.
-fn bootstrap_ip(
-    link: PhysicalLink,
-    interface_id: u64,
-) -> Result<Ipv6Addr, dladm::GetMacError> {
-    let mac = Dladm::get_mac(link)?;
-    Ok(mac_to_bootstrap_ip(mac, interface_id))
-}
-
 // Deletes all state which may be left-over from a previous execution of the
 // Sled Agent.
 //
@@ -296,8 +263,8 @@ impl Agent {
             "component" => "BootstrapAgent",
         ));
         let link = config.link.clone();
-        let ip = bootstrap_ip(link.clone(), 1)?;
-        let switch_zone_bootstrap_address = bootstrap_ip(link, 2)?;
+        let ip = bootstrap_ip(&link, 1)?;
+        let switch_zone_bootstrap_address = bootstrap_ip(&link, 2)?;
 
         // We expect this directory to exist - ensure that it does, before any
         // subsequent operations which may write configs here.
@@ -417,8 +384,7 @@ impl Agent {
         let underlay_etherstub_vnic =
             underlay_etherstub_vnic(&underlay_etherstub)?;
         let bootstrap_etherstub = bootstrap_etherstub()?;
-        let link = self.config.link.clone();
-        let switch_zone_bootstrap_address = bootstrap_ip(link, 2)?;
+        let switch_zone_bootstrap_address = bootstrap_ip(&self.config.link, 2)?;
         let hardware_monitor = HardwareMonitor::new(
             &self.log,
             &self.sled_config,
@@ -992,16 +958,6 @@ mod tests {
     use super::*;
     use macaddr::MacAddr6;
     use uuid::Uuid;
-
-    #[test]
-    fn test_mac_to_bootstrap_ip() {
-        let mac = MacAddr("a8:40:25:10:00:01".parse::<MacAddr6>().unwrap());
-
-        assert_eq!(
-            mac_to_bootstrap_ip(mac, 1),
-            "fdb0:a840:2510:1::1".parse::<Ipv6Addr>().unwrap(),
-        );
-    }
 
     #[test]
     fn persistent_sled_agent_request_serialization_round_trips() {
