+++ sled-hardware/Cargo.toml
@@ -11,7 +11,9 @@ cfg-if.workspace = true
 futures.workspace = true
 illumos-utils.workspace = true
 libc.workspace = true
+macaddr.workspace = true
 nexus-client.workspace = true
+omicron-common.workspace = true
 schemars.workspace = true
 serde.workspace = true
 slog.workspace = true
