+++ sled-hardware/src/underlay.rs
@@ -7,11 +7,20 @@
 use crate::is_gimlet;
 use illumos_utils::addrobj;
 use illumos_utils::addrobj::AddrObject;
+use illumos_utils::dladm;
 use illumos_utils::dladm::Dladm;
 use illumos_utils::dladm::FindPhysicalLinkError;
 use illumos_utils::dladm::PhysicalLink;
 use illumos_utils::dladm::CHELSIO_LINK_PREFIX;
 use illumos_utils::zone::Zones;
+use omicron_common::api::external::MacAddr;
+use std::net::Ipv6Addr;
+
+/// Initial octet of IPv6 for bootstrap addresses.
+pub const BOOTSTRAP_PREFIX: u16 = 0xfdb0;
+
+/// IPv6 prefix mask for bootstrap addresses.
+pub const BOOTSTRAP_MASK: u8 = 64;
 
 // Names of VNICs used as underlay devices for the xde driver.
 //
@@ -37,17 +46,12 @@ pub enum Error {
     FindLinks(#[from] FindPhysicalLinkError),
 }
 
+/// Convenience function that calls
+/// `ensure_links_have_global_zone_link_local_v6_addresses()` with the links
+/// returned by `find_chelsio_links()`.
 pub fn find_nics() -> Result<Vec<AddrObject>, Error> {
     let underlay_nics = find_chelsio_links()?;
-
-    let mut addr_objs = Vec::with_capacity(underlay_nics.len());
-    for nic in underlay_nics {
-        let addrobj = AddrObject::link_local(&nic.0)?;
-        Zones::ensure_has_link_local_v6_address(None, &addrobj)?;
-        addr_objs.push(addrobj);
-    }
-
-    Ok(addr_objs)
+    ensure_links_have_global_zone_link_local_v6_addresses(&underlay_nics)
 }
 
 /// Return the Chelsio links on the system.
@@ -68,5 +72,63 @@ pub fn find_chelsio_links() -> Result<Vec<PhysicalLink>, Error> {
             .into_iter()
             .map(|name| PhysicalLink(name.to_string()))
             .collect())
+    }
+}
+
+/// Ensure each of the `PhysicalLink`s has a link local IPv6 address in the
+/// global zone.
+pub fn ensure_links_have_global_zone_link_local_v6_addresses(
+    links: &[PhysicalLink],
+) -> Result<Vec<AddrObject>, Error> {
+    let mut addr_objs = Vec::with_capacity(links.len());
+
+    for link in links {
+        let addrobj = AddrObject::link_local(&link.0)?;
+        Zones::ensure_has_link_local_v6_address(None, &addrobj)?;
+        addr_objs.push(addrobj);
+    }
+
+    Ok(addr_objs)
+}
+
+// TODO(https://github.com/oxidecomputer/omicron/issues/945): This address
+// could be randomly generated when it no longer needs to be durable.
+pub fn bootstrap_ip(
+    link: &PhysicalLink,
+    interface_id: u64,
+) -> Result<Ipv6Addr, dladm::GetMacError> {
+    let mac = Dladm::get_mac(link)?;
+    Ok(mac_to_bootstrap_ip(mac, interface_id))
+}
+
+fn mac_to_bootstrap_ip(mac: MacAddr, interface_id: u64) -> Ipv6Addr {
+    let mac_bytes = mac.into_array();
+    assert_eq!(6, mac_bytes.len());
+
+    Ipv6Addr::new(
+        BOOTSTRAP_PREFIX,
+        ((mac_bytes[0] as u16) << 8) | mac_bytes[1] as u16,
+        ((mac_bytes[2] as u16) << 8) | mac_bytes[3] as u16,
+        ((mac_bytes[4] as u16) << 8) | mac_bytes[5] as u16,
+        (interface_id >> 48 & 0xffff).try_into().unwrap(),
+        (interface_id >> 32 & 0xffff).try_into().unwrap(),
+        (interface_id >> 16 & 0xffff).try_into().unwrap(),
+        (interface_id & 0xfff).try_into().unwrap(),
+    )
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use macaddr::MacAddr6;
+
+    #[test]
+    fn test_mac_to_bootstrap_ip() {
+        let mac = MacAddr("a8:40:25:10:00:01".parse::<MacAddr6>().unwrap());
+
+        assert_eq!(
+            mac_to_bootstrap_ip(mac, 1),
+            "fdb0:a840:2510:1::1".parse::<Ipv6Addr>().unwrap(),
+        );
     }
 }
